moved editPopup to profiles folder
